split out http client and clean up tests
